requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
iniconfig==2.0.0
packaging==23.2
pip==23.3.2
pluggy==1.3.0
pytest==7.4.3
requests==2.31.0
setuptools==60.2.0
urllib3==2.1.0
wheel==0.37.1

Faker~=22.1.0
